pyproject.toml
[project]
name = "dijkfood-cloud-comp-a1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.42.79",
    "faker>=40.12.0",
    "osmnx>=2.1.0",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
]
